refactor: move all string producers into renderer.ts; display.ts is I/O only
Renderer = produces strings (no I/O, no config access).
Display = manages terminal I/O (single stdout doorway, status bar, verbose prefix).

Moved from display.ts → renderer.ts:
- Content block types (ToolUseBlock, ToolResultBlock, ContentBlock)
- Format table types (Fmt, FmtEntry, FmtTable)
- clearBreak(verbose?) — now takes verbose param instead of reading getConfig()
- fmtToolCall, fmtSubagentType helpers
- All format dispatch tables (ASSISTANT_BLOCK_FMT, USER_BLOCK_FMT,
  TOOL_CALL_FMT, TOOL_RESULT_FMT, TOOL_ERROR_FMT, SYSTEM_FMT,
  MESSAGE_FMT, FOREMAN_MESSAGE_FMT) and their private helpers

display.ts now just:
- Imports FmtTables from renderer.ts and uses them via this.resolve()
- Delegates clearBreak() → renderer.clearBreak(this.config.verbose)
- Delegates effectiveWidth() → style.effectiveWidth(fallback, this.config.verbose)
- Handles all stdout writes (print, _printLine, status bar lifecycle)
